+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add `len` & `is_empty` functions
+
 # 0.17.0
 * (breaking) removed `Toast::font(font: FontId)`, this can now be done by using `egui::widget_text::RichText` and `RichText::font`. [#34]
 * Added support for `egui::widget_text::WidgetText` in Toasts, this allows for more customization of the text in the toast. [#34]